src/main/java/tapkomet/springframework/recipe/converters/UnitOfMeasureCommandToUnitOfMeasure.java
package tapkomet.springframework.recipe.converters;

import lombok.Synchronized;
import org.springframework.core.convert.converter.Converter;
import org.springframework.lang.Nullable;
import org.springframework.stereotype.Component;
import tapkomet.springframework.recipe.commands.UnitOfMeasureCommand;
import tapkomet.springframework.recipe.domain.UnitOfMeasure;

/**
 * Created by Tapkomet on 2/12/2020
 */
@Component
    public class UnitOfMeasureCommandToUnitOfMeasure implements Converter<UnitOfMeasureCommand, UnitOfMeasure> {

    @Synchronized
    @Nullable
    @Override
    public UnitOfMeasure convert(UnitOfMeasureCommand source) {
        if (source == null) {
            return null;
        }

        return UnitOfMeasure.builder().id(source.getId()).name(source.getName()).build();
    }
}
